pyproject.toml
[tool.poetry]
name = "themis"
version = "0.1.0"
description = ""
authors = ["zhikh <zhikh.k@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-levenshtein = "^0.27.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
